+++ client/src/components/Panels/InfoPanel/InfoPanel.module.css
@@ -1,0 +1,5 @@
+.mainPanel {
+  text-align: center;
+  padding: 13vh;
+  height: 84vh;
+}
