+++ client/src/components/right-section/RightSection.module.css
@@ -46,9 +46,3 @@
   height: 48px;
   width: 66px;
 }
-
-.mainPanel {
-  text-align: center;
-  padding: 13vh;
-  height: 84vh;
-}
